+++ auth/lib/authenticators/basic_auth.dart
@@ -63,7 +63,7 @@ class BasicAuth<UserModel extends PasswordUser>
 
     if (subject == null) throw UnauthorizedException.subjectNotFound;
 
-    if (!hasher.verify(password, subject.password))
+    if (!hasher.verify(subject.password, password))
       throw UnauthorizedException.invalidPassword;
 
     if (manageSession is bool && manageSession) {