+++ auth/lib/authenticators/json_auth.dart
@@ -54,7 +54,7 @@ class JsonAuth<UserModel extends PasswordUser>
 
     if (subject == null) throw UnauthorizedException.subjectNotFound;
 
-    if (!hasher.verify(password, subject.password))
+    if (!hasher.verify(subject.password, password))
       throw UnauthorizedException.invalidPassword;
 
     if (manageSession is bool && manageSession) {